create a base class for trainers, implement a trainer that uses the F…

+++ ModelTraining/BitcoinPriceForecastingTaining/Entities/CoinHistoricalData.cs
@@ -5,12 +5,12 @@ namespace BitcoinPriceForecastingTaining.Entities
     public class CoinHistoricalData
     {
         [JsonPropertyName("prices")]
-        public List<List<double>> Prices { get; set; } = null!;
+        public List<List<float>> Prices { get; set; } = null!;
 
         [JsonPropertyName("market_caps")]
-        public List<List<double>> MarketCaps { get; set; } = null!;
+        public List<List<float>> MarketCaps { get; set; } = null!;
 
         [JsonPropertyName("total_volumes")]
-        public List<List<double>> TotalVolumes { get; set; } = null!;
+        public List<List<float>> TotalVolumes { get; set; } = null!;
     }
 }
